feat: add multi-app block, change style (#335)
* fix: ci on dev (#329)

* feat: add multi-app block, change style

* feat: add app manager block

* feat: add multi-app manager plugin

* refactor: refactor multi-app plugin

* feat: add multi-app pane

* feat: add media card

* feat: multi-app, view multi-app pane

* feat: multi-app, color and icon picker

* feat: multi-app, add app list

* feat: multi-app, add app list

* feat: multi-app

---------
