+++ packages/client/src/modules/blocks/data-blocks/table/hooks/useTableBlockProps.tsx
@@ -14,9 +14,10 @@ export const useTableBlockProps = () => {
   const fieldSchema = useFieldSchema();
 
   const ctx = useTableBlockContext();
-  const globalSort = fieldSchema.parent?.['x-decorator-props']?.['dragSort'] === true
-    ? fieldSchema.parent?.['x-decorator-props']?.['dragSortBy']
-    : undefined;
+  const globalSort =
+    fieldSchema.parent?.['x-decorator-props']?.['dragSort'] === true
+      ? fieldSchema.parent?.['x-decorator-props']?.['dragSortBy']
+      : undefined;
   const { getDataBlocks } = useFilterBlock();
   const isLoading = ctx?.service?.loading;
   const params = ctx?.service?.params;
@@ -77,7 +78,7 @@ export const useTableBlockProps = () => {
             ? [sorter.field]
             : [`-${sorter.field}`]
           : parentSortReal;
-      
+
       ctx.service.run(
         { ...ctx.service.params[0], page: current, pageSize, sort },
         { filters: ctx.service.params[1]?.filters },
