+++ packages/client/src/schema-component/antd/media-card/index.ts
@@ -1,0 +1,1 @@
+export * from './MediaCard';
