+++ packages/module-multi-app/package.json
@@ -12,12 +12,12 @@
     "build": "tachybase-build --no-dts @tachybase/module-multi-app"
   },
   "dependencies": {
+    "antd": "5.22.5",
     "qs": "^6.13.1"
   },
   "devDependencies": {
     "@ant-design/icons": "^5.5.2",
     "@tachybase/schema": "workspace:*",
-    "antd": "5.22.5",
     "async-mutex": "^0.3.2",
     "lodash": "4.17.21",
     "mysql2": "^3.11.5",
