+++ packages/module-multi-app/src/client/base/MultiAppManagerProvider.style.ts
