+++ packages/module-multi-app/src/client/base/collections/collectionMultiApp.ts
@@ -1,0 +1,74 @@
+import { i18nText } from '../../locale';
+
+export const collectionMultiApp = {
+  name: 'applications',
+  filterTargetKey: 'name',
+  targetKey: 'name',
+  fields: [
+    {
+      type: 'uid',
+      name: 'name',
+      primaryKey: true,
+      prefix: 'a',
+      interface: 'input',
+      uiSchema: {
+        type: 'string',
+        title: i18nText('App ID'),
+        required: true,
+        'x-component': 'Input',
+        'x-validator': 'uid',
+      },
+    },
+    {
+      type: 'string',
+      name: 'displayName',
+      interface: 'input',
+      uiSchema: {
+        type: 'string',
+        title: i18nText('App display name'),
+        required: true,
+        'x-component': 'Input',
+      },
+    },
+    {
+      type: 'string',
+      name: 'pinned',
+      interface: 'checkbox',
+      uiSchema: {
+        type: 'boolean',
+        'x-content': i18nText('Pin to menu'),
+        'x-component': 'Checkbox',
+      },
+    },
+    {
+      type: 'string',
+      name: 'status',
+      interface: 'radioGroup',
+      defaultValue: 'pending',
+      uiSchema: {
+        type: 'string',
+        title: i18nText('App status'),
+        enum: [
+          { label: 'Initializing', value: 'initializing' },
+          { label: 'Initialized', value: 'initialized' },
+          { label: 'Running', value: 'running' },
+          { label: 'Commanding', value: 'commanding' },
+          { label: 'Stopped', value: 'stopped' },
+          { label: 'Error', value: 'error' },
+          { label: 'Not found', value: 'not_found' },
+        ],
+        'x-component': 'Radio.Group',
+      },
+    },
+    {
+      type: 'string',
+      name: 'isTemplate',
+      interface: 'checkbox',
+      uiSchema: {
+        type: 'boolean',
+        'x-content': i18nText('Is template'),
+        'x-component': 'Checkbox',
+      },
+    },
+  ],
+};
