+++ packages/module-multi-app/src/client/index.ts
@@ -1,0 +1,4 @@
+export { default } from './plugin';
+
+export { formSchema } from './base/schemas/schemaForm';
+export { tableActionColumnSchema } from './base/schemas/schemaTableActionColumn';
