+++ packages/module-multi-app/src/client/locale.tsx
