+++ packages/module-multi-app/src/client/system/hooks/useCreateDatabaseConnectionAction.tsx
@@ -1,0 +1,28 @@
+import { useActionContext, useDataBlockRequest, useDataBlockResource } from '@tachybase/client';
+import { useField, useForm } from '@tachybase/schema';
+
+export const useCreateDatabaseConnectionAction = () => {
+  const form = useForm();
+  const field = useField();
+  const ctx = useActionContext();
+  const service = useDataBlockRequest();
+  const resource = useDataBlockResource();
+  return {
+    async onClick() {
+      try {
+        await form.submit();
+        field.data = field.data || {};
+        field.data.loading = true;
+        await resource.create({ values: form.values });
+        ctx.setVisible(false);
+        await form.reset();
+        field.data.loading = false;
+        service.refresh();
+      } catch (error) {
+        if (field.data) {
+          field.data.loading = false;
+        }
+      }
+    },
+  };
+};
