+++ packages/module-multi-app/src/client/system/hooks/useMultiAppUpdateAction.tsx
@@ -1,0 +1,36 @@
+import {
+  useActionContext,
+  useCollectionRecordData,
+  useDataBlockRequest,
+  useDataBlockResource,
+} from '@tachybase/client';
+import { useField, useForm } from '@tachybase/schema';
+
+export const useMultiAppUpdateAction = (actionCallback?: (key: string, values: any) => void) => {
+  const field = useField();
+  const form = useForm();
+  const ctx = useActionContext();
+  // const { refresh } = useDataBlockRequest();
+  const service = useDataBlockRequest();
+  const resource = useDataBlockResource();
+  const record = useCollectionRecordData();
+  const filterByTk = record?.name;
+  return {
+    async onClick() {
+      await form.submit();
+      field.data = field.data || {};
+      field.data.loading = true;
+      try {
+        const res = await resource.update({ filterByTk, values: form.values });
+        ctx.setVisible(false);
+        await form.reset();
+        service.refresh();
+        actionCallback?.(filterByTk, res?.data?.data);
+      } catch (e) {
+        console.error(e);
+      } finally {
+        field.data.loading = false;
+      }
+    },
+  };
+};
