+++ packages/module-multi-app/src/client/system/hooks/useRouteUrl.ts
@@ -1,0 +1,9 @@
+import { useApp } from '@tachybase/client';
+
+export function useRouteUrl({ name, cname }) {
+  const app = useApp();
+  if (cname) {
+    return `//${cname}`;
+  }
+  return app.getRouteUrl(`/apps/${name}/admin/`);
+}
