+++ packages/module-multi-app/src/client/usage/AppList.style.ts
@@ -1,0 +1,11 @@
+import { createStyles } from '@tachybase/client';
+
+export const useStyles = createStyles(({ css, token }) => {
+  return {
+    appListStyle: css`
+      display: flex;
+      flex-wrap: wrap;
+      gap: 16px;
+    `,
+  };
+});
