+++ packages/module-multi-app/src/client/usage/MultiAppBlockInitializer.schema.ts
@@ -1,0 +1,15 @@
+export const schemaMultiAppBlockInitializer = {
+  type: 'void',
+  'x-component': 'CardItem',
+  'x-toolbar': 'BlockSchemaToolbar',
+  'x-settings': 'blockSettings:table',
+  properties: {
+    app: {
+      type: 'void',
+      'x-component': 'AppManager',
+      'x-component-props': {
+        admin: false,
+      },
+    },
+  },
+};
