+++ packages/module-multi-app/src/client/usage/ShowMultiAppBlockInitializer.schema.ts
@@ -1,0 +1,15 @@
+export const schemaShowMultiAppBlockInitializer = {
+  type: 'void',
+  'x-component': 'CardItem',
+  'x-toolbar': 'BlockSchemaToolbar',
+  'x-settings': 'blockSettings:table',
+  properties: {
+    app: {
+      type: 'void',
+      'x-component': 'ViewMultiAppPane',
+      'x-component-props': {
+        admin: false,
+      },
+    },
+  },
+};
