+++ packages/module-multi-app/src/client/usage/ViewMultiAppPane.tsx
@@ -1,0 +1,20 @@
+import { SchemaComponent, useCurrentUserContext } from '@tachybase/client';
+
+import { AppList } from './AppList';
+import { schemaViewMultiAppPane } from './ViewMultiAppPane.schema';
+
+export const ViewMultiAppPane = () => {
+  const currentUser = useCurrentUserContext();
+  const userId = currentUser?.data?.data?.id;
+  return (
+    <SchemaComponent
+      schema={schemaViewMultiAppPane}
+      scope={{
+        userId,
+      }}
+      components={{
+        AppList,
+      }}
+    />
+  );
+};
