+++ packages/module-multi-app/src/locale/en-US.json
@@ -2,6 +2,7 @@
   "All sub-applications have stopped serving, unless you reopen them": "All sub-applications have stopped serving, unless you reopen them",
   "App ID": "App ID",
   "App display name": "App display name",
+  "App is not running": "App is not running",
   "App status": "App status",
   "Applications": "Applications",
   "Auto start": "Auto start",
@@ -9,6 +10,7 @@
   "Custom domain prefix": "Custom domain prefix",
   "Is template": "Is template",
   "Manage applications": "Manage applications",
+  "Multi-app display": "Multi-app display",
   "Multi-app manager": "Multi-app manager",
   "Pin to menu": "Pin to menu",
   "Preset": "Preset",
