+++ packages/module-multi-app/src/locale/zh-CN.json
@@ -4,6 +4,7 @@
   "App already started": "应用已启动",
   "App already stopped": "应用已关闭",
   "App display name": "应用名称",
+  "App is not running": "应用未启动, 需要先在<系统-多应用管理>里启动应用",
   "App status": "应用状态",
   "Applications": "应用",
   "Auto start": "自动启动",
@@ -11,6 +12,7 @@
   "Custom domain prefix": "自定义域名前缀",
   "Is template": "是否模板",
   "Manage applications": "管理应用",
+  "Multi-app display": "多应用展示",
   "Multi-app manager": "多应用管理",
   "Pin to menu": "在菜单上显示",
   "Preset": "预设",
