+++ packages/module-multi-app/src/server/collections/applications.ts
@@ -42,6 +42,12 @@ export default defineCollection({
     {
       type: 'string',
       name: 'icon',
+      interface: 'icon',
+    },
+    {
+      type: 'string',
+      name: 'color',
+      interface: 'color',
     },
     {
       type: 'string',
